refactor: reorganize hooks and components for academic schedule and pensum management
