+++ src/app/academic-schedule/hooks/useAcademicSchedule.ts
