+++ src/app/academic-schedule/hooks/useAcademicScheduleCreate.ts
